+++ test/tod/shift_test.rb
@@ -2,6 +2,16 @@ require File.expand_path(File.join(File.dirname(__FILE__),'..','test_helper'))
 require 'active_support/time'
 
 describe "Shift" do
+  describe "#initialize" do
+    it "parses bounds" do
+      shift = Tod::Shift.new Tod::TimeOfDay.new(8), Tod::TimeOfDay.new(10), false
+      refute shift.exclude_end?
+
+      shift = Tod::Shift.new Tod::TimeOfDay.new(8), Tod::TimeOfDay.new(10), true
+      assert shift.exclude_end?
+    end
+  end
+
   describe "#duration" do
     it "returns correct duration when first time is lower than the second one" do
       duration_expected = 4 * 60 * 60 + 30 * 60 + 30 # 4 hours, 30 min and 30 sec later
@@ -91,6 +101,24 @@ describe "Shift" do
       value = Tod::TimeOfDay.new 8
       shift = Tod::Shift.new tod1, tod2
       refute shift.include?(value)
+    end
+
+    # |------------------------|--------T1-----T2V-------|------------------------|
+    it "is false when value is equal to second ToD and Shift is ending exclusive" do
+      tod1  = Tod::TimeOfDay.new 10
+      tod2  = Tod::TimeOfDay.new 16
+      value = Tod::TimeOfDay.new 16
+      shift = Tod::Shift.new tod1, tod2, true
+      refute shift.include?(value)
+    end
+
+    # |------------------------|--------T1-----T2V-------|------------------------|
+    it "is true when value is equal to second ToD and Shift is ending inclusive" do
+      tod1  = Tod::TimeOfDay.new 10
+      tod2  = Tod::TimeOfDay.new 16
+      value = Tod::TimeOfDay.new 16
+      shift = Tod::Shift.new tod1, tod2, false
+      assert shift.include?(value)
     end
   end
 end