+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Shift now supports exclusive ends like a Ruby range
 * Do not pollute global namespace
 
 # 1.5.0 (January 15, 2015)